fixed deserealization

+++ extern-api-java-sdk-integrational-tests/src/test/java/ru/kontur/extern_api/sdk/RelatedDocumentsIT.java
@@ -318,7 +318,7 @@ class RelatedDocumentsIT {
                 .join();
 
         assertEquals(sentRelatedInventories + sentRelatedLetters,
-                demandAttachment.getDescription().getRelatedDocflowsCount());
+                demandAttachment.getDescription().getRelatedDocflowsCount().longValue());
     }
 
     private CompletableFuture<Docflow> sendRelatedLetter(DemandTestData testData) {
